Refactor HTTP calls
We must always use lizmapProxy::getRemoteData instead of curl, as this method
take care of some lizmap configuration parameters like proxy and so on,
contrary to the current code.

We should not use deprecated parameters of getRemoteData.

To simplify the code, get and post methods have been replaced by
a single method doRequest.