Fix potential issues when calling WPS requests with unexpected method
Even if the request type name is checked into controllers, when we used one of the WPS requests object independently, it could be processed with the name of any methods of the object.

Only methods corresponding to the request should be able to be executed. This is why I choose the solution to prefix these methods, so no other methods can be executed.

This naming allows too to identify easily methods of request types.

The issue is related to https://github.com/3liz/lizmap-web-client/pull/2256