Fix: Refused to set unsafe header "Content-Length"
XMLHttpRequest isn't allowed to set this header, it is being set automatically by the browser.
The reason is that by manipulating this header you might be able to trick the server into accepting a second request through the same connection,
one that wouldn't go through the usual security checks - that would be a security vulnerability in the browser.

It was an old setting for old browser.

Fixes #17